requirements.txt
tensorflow==2.18.0
numpy==2.0.2
opencv-python==4.10.0.84
flask==3.1.0
pillow==11.1.0
